fix: create services package and moved db
